requirements.txt
nose >= 1.3.0
gitpython >= 2.1.0
pandas >= 0.23.0
matplotlib >= 3.0.0
scikit-learn >= 0.2.0
statsmodels >= 0.9.0
seaborn >= 0.9.0
